Factored out components from CameraActivity into AbstractCameraActivity
