+++ app/src/main/java/com/lindronics/flirapp/activities/CameraActivity.java
@@ -2,37 +2,15 @@ package com.lindronics.flirapp.activities;
 
 import android.graphics.Bitmap;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
-import android.widget.ImageView;
 import android.widget.ToggleButton;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.flir.thermalsdk.ErrorCode;
-import com.flir.thermalsdk.live.Identity;
-import com.flir.thermalsdk.live.connectivity.ConnectionStatus;
-import com.flir.thermalsdk.live.connectivity.ConnectionStatusListener;
-import com.google.gson.Gson;
-import com.lindronics.flirapp.camera.CameraHandler;
 import com.lindronics.flirapp.R;
 import com.lindronics.flirapp.camera.FrameDataHolder;
 import com.lindronics.flirapp.camera.ImageWriter;
 
-import org.jetbrains.annotations.NotNull;
-
-import java.util.concurrent.LinkedBlockingQueue;
-
-public class CameraActivity extends AppCompatActivity {
+public class CameraActivity extends AbstractCameraActivity {
 
     private static final String TAG = "CameraActivity";
-
-    private CameraHandler cameraHandler;
-
-    private ImageView rgbImage;
-    private ImageView firImage;
-
-    private LinkedBlockingQueue<FrameDataHolder> framesBuffer = new LinkedBlockingQueue<>(21);
 
     private ToggleButton cameraButton;
 
@@ -42,101 +20,29 @@ public class CameraActivity extends AppCompatActivity {
     /**
      * Executed when activity is created.
      * Get camera identity from intent and connect to camera.
-     *
      */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_camera);
+        super.onCreate(savedInstanceState);
 
-        cameraHandler = new CameraHandler();
-
-        rgbImage = findViewById(R.id.rgb_view);
-        firImage = findViewById(R.id.fir_view);
         cameraButton = findViewById(R.id.camera_button);
-
-        Bundle extras = getIntent().getExtras();
-        Gson gson = new Gson();
-
-        if (extras == null) {
-            finish();
-            return;
-        }
-
-        String identityString = extras.getString("cameraIdentity");
-        Identity cameraIdentity = gson.fromJson(identityString, Identity.class);
-        cameraHandler.connect(cameraIdentity, connectionStatusListener);
-
     }
 
-    /**
-     * Defines behaviour for when images are received
-     */
-    private final CameraHandler.StreamDataListener streamDataListener = new CameraHandler.StreamDataListener() {
+    @Override
+    public void images(FrameDataHolder dataHolder) {
+        super.images(dataHolder);
+    }
 
-        @Override
-        public void images(FrameDataHolder dataHolder) {
+    @Override
+    public void images(Bitmap firBitmap, Bitmap rgbBitmap) {
 
-            runOnUiThread(() -> {
-                firImage.setImageBitmap(dataHolder.firBitmap);
-                rgbImage.setImageBitmap(dataHolder.rgbBitmap);
-            });
+        super.images(firBitmap, rgbBitmap);
+
+        if (imageWriter != null) {
+            imageWriter.saveImages(firBitmap, rgbBitmap);
         }
-
-        @Override
-        public void images(Bitmap firBitmap, Bitmap rgbBitmap) {
-
-            try {
-                framesBuffer.put(new FrameDataHolder(firBitmap, rgbBitmap));
-            } catch (InterruptedException e) {
-                // If interrupted while waiting for adding a new item in the queue
-                Log.e(TAG, "images(), unable to add incoming images to frames buffer, exception:" + e);
-            }
-
-            runOnUiThread(() -> {
-                Log.d(TAG, "framebuffer size:" + framesBuffer.size());
-                FrameDataHolder poll = framesBuffer.poll();
-                if (poll != null) {
-                    firImage.setImageBitmap(poll.firBitmap);
-                    rgbImage.setImageBitmap(poll.rgbBitmap);
-                }
-            });
-
-            if (imageWriter != null) {
-                imageWriter.saveImages(firBitmap, rgbBitmap);
-            }
-        }
-    };
-
-
-    /**
-     * Defines behaviour vor changes of the connection status
-     */
-    private final ConnectionStatusListener connectionStatusListener = new ConnectionStatusListener() {
-        @Override
-        public void onConnectionStatusChanged(@NotNull ConnectionStatus connectionStatus, @org.jetbrains.annotations.Nullable ErrorCode errorCode) {
-            Log.d(TAG, "onConnectionStatusChanged connectionStatus:" + connectionStatus + " errorCode:" + errorCode);
-
-            runOnUiThread(() -> {
-
-                switch (connectionStatus) {
-                    case CONNECTING:
-                        break;
-                    case CONNECTED: {
-                        cameraHandler.startStream(streamDataListener);
-                    }
-                    break;
-                    case DISCONNECTING:
-                        break;
-                    case DISCONNECTED: {
-                        endCapture();
-                        finish();
-                    }
-                    break;
-                }
-            });
-        }
-    };
+    }
 
     /**
      * Event listener for starting or stopping camera capture/recording
@@ -152,13 +58,27 @@ public class CameraActivity extends AppCompatActivity {
         }
     }
 
+    /**
+     * Start capturing/recording data
+     */
     private void startCapture() {
         cameraButton.setBackground(getDrawable(R.drawable.ic_camera_capture_recording));
         imageWriter = new ImageWriter(this);
     }
 
+    /**
+     * Finish capturing/recording data
+     */
     private void endCapture() {
         cameraButton.setBackground(getDrawable(R.drawable.ic_camera_capture_ready));
         imageWriter = null;
+    }
+
+    /**
+     * End capture when disconnected
+     */
+    @Override
+    void onDisconnected() {
+        endCapture();
     }
 }
